requirements.txt
streamlit
fpdf2
Pillow
google-api-python-client
google-auth-httplib2
google-auth-oauthlib